This is the first commit of the Minolta Dimage V camera library. This is a new work, owing very little to my previous gphoto 0.4.x Minolta Dimage V driver. It support image capture, download, and deletion.
This does not support any other model of Minolta camera to the best of my
knowledge, only the now-discontinued Dimage V.

The C files are divided by function; the primary file is dimagev.c, with
subroutines of varying types divided into the rest. For example, packet.c
contains functions pertaining to the packet communication used by the camera.


git-svn-id: https://gphoto.svn.sourceforge.net/svnroot/gphoto/trunk/libgphoto2@1121 21bdcb62-e50e-0410-80e4-ae60ad24c865