git: Add gphoto m4 macros as submodule */m4m
Add git submodules for the gphoto m4 macros as a replacement for
the SVN svn:externals mechanism. We need the macros twice, namely
in m4m/ and libgphoto2_port/m4m/.

Note that the libgphoto2_port/m4m/ location is different from the
original location libgphoto2_port/m4/ in order to make the layout
more uniform.