st2205: Fixup checksum for 128x160 shuffle tables
Now that we only use 7 of the 8 tables, as the device itself seems to
not be able to properly use the 8th table, the checksum of the tables
we load has changed.

p.s.
I've such a device in question myself now and I can confirm this problem,
I guess that the issue is that the 7th table crosses a 0x4000 address
boundary, and the device screws up with paging or some such.


git-svn-id: https://gphoto.svn.sourceforge.net/svnroot/gphoto/branches/libgphoto2-2_4/libgphoto2@12829 21bdcb62-e50e-0410-80e4-ae60ad24c865