ax203: ABFS location differs between models
The ABFS (PAT) location is not constant between different model
frames, so search for it (in 0x10000 byte intervals).

While working on this I also noticed that the v3.3.x and v3.4.x firmware
ABFS is not 0x2000 bytes large, but only 0x1000 bytes, the second 0x1000
bytes is a 2nd parameter block. This patch also fixes this and fixes
delete_all to no overwrite this parameter block.


git-svn-id: https://gphoto.svn.sourceforge.net/svnroot/gphoto/branches/libgphoto2-2_4/libgphoto2@12913 21bdcb62-e50e-0410-80e4-ae60ad24c865